Add `p-queue` to control concurrency
Previously, we used `pendingUpdates` to keep track of when all work was
finished. However, this also meant that `pendingUpdates` grew with the
size of the input file.

With `p-queue`, we can use `onSizeLessThan` to prevent us from eagerly
reading in the entire file, and controlling the total queue size.

Also, it gives us `onIdle` which is a built-in way to wait until all
work has stopped before wrapping up.